backend/pyproject.toml
[tool.poetry]
name = "myportfoliospa"
version = "0.1.0"
description = ""
authors = ["maxx_PC <m.petrykin@gmx.de>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.2.6"
djangorestframework = "^3.16.1"
django-cors-headers = "^4.7.0"
drf-yasg = "^1.21.10"
psycopg2-binary = "^2.9.10"
python-dotenv = "^1.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
